+++ src/main/java/com/catkatpowered/katserver/tokenpool/KatTokenPool.java
@@ -46,8 +46,8 @@ public class KatTokenPool {
             case All -> this.tokenPool.clear();
             case Outdated -> {
                 int outdatedTime = KatServer.KatConfigAPI
-                        .<Integer>getConfig(KatConfigNodeConstants.KAT_CONFIG_TOKENPOOL_OUTDATED)
-                        .get();
+                        .<Long>getConfig(KatConfigNodeConstants.KAT_CONFIG_TOKENPOOL_OUTDATED)
+                        .get().intValue();
                 for (var item : this.tokenPool.entrySet())
                     if ((System.currentTimeMillis() - item.getValue()) >= outdatedTime)
                         this.tokenPool.remove(item.getKey());
